Depth Stencil View

+++ Engine/TableDescriptorHeap.cpp
@@ -45,13 +45,13 @@ void TableDescriptorHeap::CommitTable()
 	D3D12_GPU_DESCRIPTOR_HANDLE handle = _descHeap->GetGPUDescriptorHandleForHeapStart();
 	handle.ptr += _currentGroupIndex * _groupSize;
 	CMD_LIST->SetGraphicsRootDescriptorTable(0, handle);
-	
+
 	_currentGroupIndex++;
 }
 
 D3D12_CPU_DESCRIPTOR_HANDLE TableDescriptorHeap::GetCPUHandle(CBV_REGISTER reg)
 {
-	return GetCPUHandle(static_cast<uint32>(reg));
+	return GetCPUHandle(static_cast<uint8>(reg));
 }
 
 D3D12_CPU_DESCRIPTOR_HANDLE TableDescriptorHeap::GetCPUHandle(SRV_REGISTER reg)
